refactor and rename test classes
